Fix schema

+++ src/main/resources/schema.sql
@@ -1,5 +1,5 @@
 create table if not exists members (
-    id int NOT NULL PRIMARY KEY AUTO_INCREMENT,
+    id bigint unsigned NOT NULL PRIMARY KEY AUTO_INCREMENT,
     name varchar(256) NOT NULL,
     facebookUrl varchar(256),
     instagramUrl varchar(256),
